Remove watched/movies and watched/shows
These two API endpoints seem to represent a summary view of the history
of a Trakt collection. As such, the history endpoints are actually more
useful as they contain the actual watched history (each watch) of an
item in the colleciton rather than a summary (total watches) which is
what the watched endpoints contain.