Add trakt-restore.sh
Add restore functionality for collections and history. This is an
experimental feature that may need to be modified for additional
scenarios not yet tested. The current restore functionality has been
tested with a collection containing approximately 4000 movies and tv
shows and a history containing approximately 3000 watches.

Future efforts may add restore functionality for ratings, watchlist,
comments, followers, and following.