Do not clear history by default during restore
To prevent duplication watch history, the script cleared the entire
history before executing the restore. However, this can (obviously) be
harmful.

Now, by default, the history is NOT cleared. To clear a history before
restoring, the -c or --clear-history argument must be set to true.

This is to allow greater freedom in the restore process. Once example
of using this functionality is to manually "reset" the history of a TV
series being watched for a second time. For instance, the TV series has
already been watched and it marked as such in Trakt. To watch the TV
series again, all episodes will already be considered watched and thus
any media servers will not keep track of "next episodes." In theory,
the entire watch history of the series could be exported from Trakt
manually erased from the Trakt history, and shown as never watched in
the media server. After watching the series a second time, the watch
history of the TV series (and only the TV series) could be restored to
Trakt in an additive way. This will preserve the fact that the series
was watched twice while allowing a media server to keep trakt of "next
episodes" in a media server.

This is, of course, hypothetical and is left up to user implementation.
However, by allowing allowing the user to decide whether or not the
Trakt history is cleared during a restore, greater freedom is allowed
in making such decisions.